before aws

+++ controllers/videoControllers.js
@@ -1,5 +1,6 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import Comment from "../models/Comment";
 
 export const homeCon = async (req, res) => {
   try {
@@ -47,10 +48,13 @@ export const postUploadCon = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
+
   // To Do: Upload and save video
   console.log(newVideo);
-  console.log(path);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videos + routes.videoDetail(newVideo.id));
 };
 
@@ -59,7 +63,9 @@ export const videoDetailCon = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id)
+      .populate(`creator`)
+      .populate(`comments`);
     res.render("videodetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -72,9 +78,12 @@ export const getEditVideoCon = async (req, res) => {
     params: { id },
   } = req;
   try {
-    console.log(id);
     const video = await Video.findById(id);
-    res.render("editvideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editvideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.home);
@@ -100,9 +109,54 @@ export const deleteVideoCon = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
+};
+
+// api register video view
+export const postRegisterViewCon = async (req, res) => {
+  const {
+    params: { id },
+  } = req;
+  try {
+    const video = await Video.findById(id);
+    video.views += 1;
+    video.save();
+    res.status(200);
+  } catch (error) {
+    res.status(400);
+    res.end();
+  } finally {
+    res.end();
+  }
+};
+
+// api add comment
+export const postAddCommentCon = async (req, res) => {
+  const {
+    params: { id },
+    body: { comment },
+    user,
+  } = req;
+  try {
+    const video = await Video.findById(id);
+    const newComment = await Comment.create({
+      text: comment,
+      creator: user.id,
+    });
+    video.comments.push(newComment.id);
+    video.save();
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
 };
